Conflicts:
	src/boston/Bus/Map/main/Main.java
	src/boston/Bus/Map/main/UpdateHandler.java
	src/boston/Bus/Map/tutorials/IntroTutorial.java
	src/boston/Bus/Map/ui/BusPopupView.java

+++ src/boston/Bus/Map/provider/DatabaseAgent.java
@@ -773,7 +773,7 @@ public class DatabaseAgent implements IDatabaseAgent {
                 float lat = (float) (builder.getLatitudeAsDegrees() * Geometry.degreesToRadians);
                 float lon = (float) (builder.getLongitudeAsDegrees() * Geometry.degreesToRadians);
                 if (filterByDistance) {
-                    float distance = stop.distanceFromInMiles(lat, lon);
+                    double distance = stop.distanceFromInMiles(lat, lon);
                     if (distance < miles) {
                         routes.addAll(stop.getRoutes());
                     }
